Day 11: The Owner (Relasi Database)

+++ main.py
@@ -83,16 +83,20 @@ def create_inventory(inventory : InventorySchema,
                      ):
     # Pastikan nama modelnya benar (InventoryDB atau InventoryBD? Cek models.py)
     # Disini saya asumsikan InventoryDB yang benar.
+    #Cek ID Baranf (Validasi Manualmu)
     cek_duplikasi = db.query(models.InventoryDB).filter(models.InventoryDB.id == inventory.id).first()
     if cek_duplikasi:
         raise HTTPException(status_code=400, detail="ID already registered")
     
+
+    # SIMPAN KE DATABASE (DENGAN STEMPEL PEMILIK)
     new_inventory = models.InventoryDB(
         id = inventory.id,
         name = inventory.name,
         price = inventory.price,
         stock = inventory.stock,
-        description = inventory.description
+        description = inventory.description,
+        owner_id = current_user.id #  <-- INI KUNCI DAY 11! Ambil ID dari user yang sedang login
     )
     db.add(new_inventory)
     db.commit()
@@ -160,9 +164,7 @@ def get_inventory(db: Session = Depends(get_db),
     return db.query(models.InventoryDB).all()
 
 
-
 # ============= UPDATE =================
-
 
 
 # Perbaikan: Tambah "/" sebelum {id}
@@ -178,6 +180,15 @@ def update_inventory(id:int,
     if db_item is None:
         raise HTTPException(status_code=404, detail="Item not Found")
     
+
+    # ======== LOGIKA DAY 12: CEK KEPEMILIKAN ========
+    # jika ID memiliki barang BEDA dengan ID user yang login
+    if db_item.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail= "Not authorized to delete this item")
+
+    # --------------------------------------------
+
+
     db_item.name = inventory_update.name
     db_item.price = inventory_update.price
     db_item.stock = inventory_update.stock
@@ -199,11 +210,21 @@ def delete_item(id: int,
                 db: Session=Depends(get_db),
                 current_user: models.UserDB = Depends(get_current_user)
                 ):
+        # 1. Cari barangnya dulu
         db_item = db.query(models.InventoryDB).filter(models.InventoryDB.id == id).first()
 
+        # 2. Kalau barang gak ada
         if db_item is None:
             raise HTTPException(status_code=404, detail="Item Not Found")
         
+
+        # ======== LOGIKA DAY 12: CEK KEPEMILIKAN ========
+        # jika ID memiliki barang BEDA dengan ID user yang login
+        if db_item.owner_id != current_user.id:
+            raise HTTPException(status_code=403, detail= "Not authorized to delete this item")
+
+        # --------------------------------------------
+
         db.delete(db_item)
         db.commit()
